Fixes and adds efm, adds --no-color option
Fixes the errorformat for check and install, introduces some for show.

Also introduced a new variable (see documentation) to have output not in
color as default, which stops messing with parsing.

In the end, as long as the 'Autocomplete' function is not fixed, it's
for nothing (or at least not much).